[FIX] Error on init.sql

+++ web/mysql/init.sql
@@ -2,7 +2,7 @@ CREATE TABLE `woodytoys`.`articles` (
   `id` INT NOT NULL,
   `name` VARCHAR(255) COLLATE utf8_unicode_ci DEFAULT NULL,
   `brand` VARCHAR(255) DEFAULT NULL,
-  `price` FLOAT(0,0) DEFAULT NULL,
+  `price` FLOAT DEFAULT NULL
 );
 
 INSERT INTO `woodytoys`.`articles`
